+++ src/stopwatch.rs
@@ -190,7 +190,7 @@ impl ops::SubAssign<Duration> for Stopwatch {
 }
 
 /// Errors associated with [`Stopwatch`]
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
 pub enum Error {
     /// Called [`Stopwatch::start`] while running
     AlreadyStarted,