+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 * renamed license files
 
 ## [0.2.0] - 2022-08-07
-* added Debug impl for Stopwatch
+* added `Debug` impl for `Stopwatch`
 * improved docs organisation
 
 ## [0.1.1] - 2022-08-07